Mark _start reference as hidden
Mark the _start symbol as hidden in ldso_trampoline.cpp so that the
compiler uses an R_386_GOTOFF relocation instead of a R_386_GOT32X
relocation.  The former uses an offset from the GOT table, which is
at a location known at link time, while the latter needs to load a
value out of the GOT table, which has not yet been initialized.

Bug: 216192129
Test: 32-bit gtest
Change-Id: I2ead955b8bc1cacae6c9af06270f5b20bba39ec7