Support searching LD_LIBRARY_PATH and DT_RUNPATH for loader
Instead of putting a single relative path into the relinterp note,
put the loader name into relinterp and reuse DT_RUNPATH or
LD_LIBRARY_PATH to find the loader.  This allows for multiple
paths in DT_RUNPATH, and for overriding the path when the loader
is not at the expected path.

Bug: 242601708
Test: clang builds against musl sysroot with relinterp
Change-Id: I9238e01264b7c940643f54c92595dab496c079b4