Ignore incorrect elf architecture loaders in relinterp
In multilib systems LD_LIBRARY_PATH is sometimes set to a list that
contains both the lib32 and lib64 directories, for example when
running code that may execute 32-bit or 64-bit subprocesses or both.
Relinterp aborts searching a path list when it finds any loader that
is not loadable, preventing searching the other multilib directory.

Modify relinterp to continue searching when a file is found that is
a valid elf file but for an elf machine or class that differs from
that of the app.

Test: host-unit-tests
Change-Id: Id9b9ffc0d964565d4da0db158671e5703c87c192